refactor: Calculate tree hierarchy outside of TreeLineageView
