+++ src/components/Tabs/Lineage/constants.ts
@@ -1,0 +1,1 @@
+export const DUMMY_ROOT_NODE_ID = -1;
