+++ src/components/Tabs/Lineage/types.ts
@@ -1,7 +1,3 @@
-import type * as d3 from "d3";
-import type React from "react";
-import type { Color } from "three";
-
 export type TrackInfo = {
   id: number;
   length: number;
@@ -28,16 +24,4 @@ export type LineageDataRelationships = {
   multiparentEdges: [number, number][];
   /** A map from a track ID to its parent track IDs. */
   idToParents: Map<number, number[]>;
-};
-
-export type SharedLineageViewProps = {
-  container: React.RefObject<HTMLDivElement>;
-  data: LineageData;
-  relationships: LineageDataRelationships;
-  selectedTracks: Set<number>;
-  trackColors: Map<number, Color>;
-  colorScale: d3.ScaleSequential<string>;
-  radiusScale: d3.ScalePower<number, number>;
-  onClick?: (trackId: number) => void;
-  onHover?: (trackId: number | null) => void;
 };
